Cache parsed FeedParserDict objects
The retrieved (and cached) requests.Response object is replaced by the
output of feedparser.parse(). This gives a significant improvement in
performance (it turns out feedparser is very slow -- running under pypy
didn't give any noticeable improvements).

Obviously it is inefficient to cache and then throw away the
requets.Response object, but using cache_get is easy and the increase
in performance by caching the output of feedparser is more than worth
it.

Squashed commit of the following:

commit 82b2e02262141dd53b803616c8d05708a16fda58
Author: A. Cynic <chris@mretc.net>
Date:   Tue Aug 29 09:28:52 2017 -0600

    Cache parsed feed without clobbering expire_dt

    This required adding a new replace_data() method to ShelfCache.

commit 64273624e82cf06d125e8d37ea47c4ee6d7b99ea
Author: A. Cynic <chris@mretc.net>
Date:   Tue Aug 29 09:26:01 2017 -0600

    Specify version of shelfcache >= 1.40

    The update_data() method was introduced in version 1.4.0

commit 7d2b1cf5a1a383f3ad87e13abccd99675707a18f
Author: A. Cynic <chris@mretc.net>
Date:   Mon Aug 28 20:38:33 2017 -0600

    Add tests for cached FeedParserDicts

commit 63c4c92ed7f21da37a50e657b2bbd70436adbfe2
Author: A. Cynic <chris@mretc.net>
Date:   Mon Aug 28 17:50:05 2017 -0600

    Only try to delete test db if it exists

commit 3777215cbc47e2a6977342838326be2dfca91c16
Author: A. Cynic <chris@mretc.net>
Date:   Mon Aug 28 15:32:33 2017 -0600

    First stab at caching parsed feed

+++ test/integration/test_feedmixer_api.py
@@ -23,7 +23,8 @@ class FMTestCase(testing.TestCase):
         self.app = feedmixer_api.wsgi_app(db_path=TESTDB)
 
     def tearDown(self):
-        os.remove(TESTDB)
+        if os.path.exists(TESTDB):
+            os.remove(TESTDB)
 
     def get_results_errors(self, path='/', qs=''):
         """
